Make a Key with make_keys! macro
We need a macro to get a variadic-ish function in Rust. This
implementation should let us make a Key from an expression like

```
let k: Key = make_keys!(v1, v2, v3);
```

Where Key has a From implementation for any of the values passed.

This is convenient because we need to concatenate keys, but it isn't really a
strong or deterministic encoding.